+++ src/mqtt/mqtt_dataUser.h
@@ -5,6 +5,13 @@
 int uBasicUnsubscribe(const char *uTopic);
 int uBasicSubscribe(const char *uTopic);
 int uBasicCloudPublish(const char *payload, const char *uTopic);
+int uCustomPublish(const char *payload, const char *uTopic);
+int uCustomSubscribe(const char *uTopic);
+int uCustomUnsubscribe(const char *uTopic);
+int uStorageServiceStore(const char *token, const char *name, float value);
+int uStorageServiceStore(const char *token, const char *name, float value, const char *callbackTopic);
+int uSmsServiceSend(const char *token, const char *receiver, const char *content);
+int uSmsServiceSend(const char *token, const char *receiver, const char *content, const char *callbackTopic);
 int uCustomSubscribe(const char *uTopic);
 int uCustomUnsubscribe(const char *uTopic);